Pass status filter presets to consumer
