+++ projects/js-packages/components/components/threats-data-views/index.tsx
@@ -1,4 +1,4 @@
-import { getThreatType, type Threat } from '@automattic/jetpack-scan';
+import { getThreatType, type Threat, type ThreatStatus } from '@automattic/jetpack-scan';
 import {
 	type Action,
 	type ActionButton,
@@ -14,7 +14,7 @@ import {
 import { dateI18n } from '@wordpress/date';
 import { __ } from '@wordpress/i18n';
 import { Icon } from '@wordpress/icons';
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useMemo, useState, useEffect } from 'react';
 import Badge from '../badge';
 import ThreatFixerButton from '../threat-fixer-button';
 import ThreatSeverityBadge from '../threat-severity-badge';
@@ -55,6 +55,7 @@ import ThreatsStatusToggleGroupControl from './threats-status-toggle-group-contr
  * @param {Function} props.isThreatEligibleForFix      - Function to determine if a threat is eligible for fixing.
  * @param {Function} props.isThreatEligibleForIgnore   - Function to determine if a threat is eligible for ignoring.
  * @param {Function} props.isThreatEligibleForUnignore - Function to determine if a threat is eligible for unignoring.
+ * @param {Function} props.onStatusFilterChange        - Callback function run when the status filter changes.
  *
  * @return {JSX.Element} The ThreatsDataViews component.
  */
@@ -68,6 +69,7 @@ export default function ThreatsDataViews( {
 	onFixThreats,
 	onIgnoreThreats,
 	onUnignoreThreats,
+	onStatusFilterChange,
 }: {
 	data: Threat[];
 	filters?: Filter[];
@@ -78,6 +80,7 @@ export default function ThreatsDataViews( {
 	onFixThreats?: ( threats: Threat[] ) => void;
 	onIgnoreThreats?: ActionButton< Threat >[ 'callback' ];
 	onUnignoreThreats?: ActionButton< Threat >[ 'callback' ];
+	onStatusFilterChange?: ( newStatus: 'active' | 'historic' | null ) => void;
 } ): JSX.Element {
 	const baseView = {
 		sort: {
@@ -476,6 +479,33 @@ export default function ThreatsDataViews( {
 	] );
 
 	/**
+	 * Memoized function to determine if a status filter is selected.
+	 *
+	 * @param {Array} threatStatuses - List of threat statuses.
+	 */
+	const isStatusFilterSelected = useMemo(
+		() => ( threatStatuses: ThreatStatus[] ) =>
+			view.filters.some(
+				filter =>
+					filter.field === 'status' &&
+					Array.isArray( filter.value ) &&
+					filter.value.length === threatStatuses.length &&
+					threatStatuses.every( threatStatus => filter.value.includes( threatStatus ) )
+			),
+		[ view.filters ]
+	);
+
+	const selectedStatusFilter = useMemo( () => {
+		if ( isStatusFilterSelected( [ 'current' ] ) ) {
+			return 'active' as const;
+		}
+		if ( isStatusFilterSelected( [ 'fixed', 'ignored' ] ) ) {
+			return 'historic' as const;
+		}
+		return null;
+	}, [ isStatusFilterSelected ] );
+
+	/**
 	 * Apply the view settings (i.e. filters, sorting, pagination) to the dataset.
 	 *
 	 * @see https://github.com/WordPress/gutenberg/blob/trunk/packages/dataviews/src/filter-and-sort-data-view.ts
@@ -500,6 +530,11 @@ export default function ThreatsDataViews( {
 	 */
 	const getItemId = useCallback( ( item: Threat ) => item.id.toString(), [] );
 
+	// Notify the consumer whenever the selectedStatusFilter changes
+	useEffect( () => {
+		onStatusFilterChange?.( selectedStatusFilter );
+	}, [ selectedStatusFilter, onStatusFilterChange ] );
+
 	return (
 		<DataViews
 			actions={ actions }
@@ -516,6 +551,7 @@ export default function ThreatsDataViews( {
 					data={ data }
 					view={ view }
 					onChangeView={ onChangeView }
+					selectedStatusFilter={ selectedStatusFilter }
 				/>
 			}
 		/>
